1859 add search cli Extract SearchEngine from SearchDialog Relates to #1859
